% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accdpx.R
\name{atom.dpx}
\alias{atom.dpx}
\title{Atom Depth Analysis}
\usage{
atom.dpx(pdb)
}
\arguments{
\item{pdb}{is either a PDB id, or the path to a pdb file.}
}
\value{
A dataframe with the computed depths.
}
\description{
Computes the depth from the surface for each protein's atom.
}
\details{
This function computes the depth, defined as the distance in angstroms between the target atom and the closest atom on the protein surface. When the protein is composed of several subunits, the calculations are made for both, the atom being part of the complex, and the atom being only part of the polypeptide chain to which it belongs.
}
\examples{
\dontrun{atom.dpx('1cll')}
}
\references{
Pintar et al. 2003. Bioinformatics 19:313-314 (PMID: 12538266)
}
\seealso{
res.dpx(), acc.dssp(), str.part()
}
\author{
Juan Carlos Aledo
}
