% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non-Exportable-Functions.R
\name{.get.url}
\alias{.get.url}
\title{Get Web Resource}
\usage{
.get.url(url, n_tries = 3)
}
\arguments{
\item{url}{url to be reached.}

\item{n_tries, }{number of tries.}
}
\value{
Returns the response or an error message.
}
\description{
Gets a web resource.
}
\keyword{internal.}
