% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssp.R
\name{download.dssp}
\alias{download.dssp}
\title{Download a Precomputed DSSP File}
\usage{
download.dssp(id, destfile = "./")
}
\arguments{
\item{id}{the 4-character identifier of the PDB structure.}

\item{destfile}{a character string with the path to the directory in which you want to save the downloaded file.}
}
\value{
Returns a precomputed dssp file that is saved at the location indicated by the user.
}
\description{
Downloads the precomputed DSSP file of the chosen PDB structure.
}
\details{
The database selected by default is 'pdb_redo', which corresponds to fully optimised structure models. If the choosen pdb structure has not a precomputed file, the function will resort to the database 'pdb'. To download these precomputed DSSP files we use the command 'rsync'. If your OS experiences problems dealing with that command, you may consider the alternative of using the function compute.dssp.
}
\examples{
download.dssp('3cwm')
}
\references{
Touw et al (2015) Nucl. Ac. Res. 43(Database issue): D364-D368 (PMID: 25352545).
}
\seealso{
parse.dssp(), compute.dssp(), mkdssp() and acc.dssp()
}
\author{
Juan Carlos Aledo
}
