% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmplot.R
\name{find.aaindex}
\alias{find.aaindex}
\title{Find the Amino Acid Indexes}
\usage{
find.aaindex(word)
}
\arguments{
\item{word}{a character string for the key-word of interest.}
}
\value{
The number and ID of the indexes matching the requested word
}
\description{
Finds amino acid indexes.
}
\examples{
find.aaindex('mutability')
find.aaindex('Kyte-Doolittle')
}
\references{
https://www.genome.jp/aaindex/AAindex/list_of_indices
}
\seealso{
ptm.plot()
}
\author{
Juan Carlos Aledo
}
