% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server_photos.R
\name{get_server_photos}
\alias{get_server_photos}
\title{Take some ptm numbers and find the photos}
\usage{
get_server_photos(ptm, results_folder = tempdir())
}
\arguments{
\item{ptm}{int a ptm number}

\item{results_folder}{chr destination folder}
}
\value{

}
\description{
Take some ptm numbers and find the photos
}
\examples{
### Not run:
#get_server_photos(5)
## End(Not run)
}
