% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterlist.R
\name{masterlist}
\alias{masterlist}
\title{Opens PTM Masterlist}
\usage{
masterlist()
}
\value{
dataframe
}
\description{
- Pulls from the PTM Masterlist Google Sheet and reads the file
- *note* you must have been granted access on your gmail before using
}
\examples{
masterlist()
}
