% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterlist_update_seq.R
\name{masterlist_update_seq}
\alias{masterlist_update_seq}
\title{Updates the sequence column of the masterlist}
\usage{
masterlist_update_seq(test = F)
}
\description{
searches for the primers: psbar2, psbaf, gwsfn, gwsrx, rbclrevnew, f57 
must have access to the lab server and [google sheet](https://docs.google.com/spreadsheets/d/1vzVIT5gjQ0yCGwbAyqB4f8pltOkLfWz27TIFTxnm9hk/edit#gid=0)
it takes a while to run because we need to go through the server to find all the sequences
}
