% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non-Exportable-Functions.R
\name{msa}
\alias{msa}
\title{Multiple Sequence Alignment}
\usage{
msa(sequences, ids = names(sequences), sfile = FALSE)
}
\arguments{
\item{sequences}{vector containing the sequences.}

\item{ids}{vector containing the sequences' ids.}

\item{sfile}{path to the file where the fasta alignmet should be saved, if any.}
}
\value{
Returns a list of four elements. The first one ($seq) provides the sequences analyzed, the second element ($ids) retuns the identifiers, ther third element ($aln) privides the alignment in fasta format and the fourth element ($ali) gives the alignment in matricial format.
}
\description{
Aligns multiple protein sequences
}
\keyword{internal.}
