% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmScan.R
\name{ni.scan}
\alias{ni.scan}
\title{Scan a Protein in Search of Nitration Sites}
\usage{
ni.scan(up_id, db = 'all')
}
\arguments{
\item{up_id}{a character string corresponding to the UniProt ID.}

\item{db}{the database where to search. It should be one among 'PSP', 'dbPTM', 'all'.}
}
\value{
Returns a dataframe where each row corresponds to a modified residue.
}
\description{
Scans the indicated protein in search of nitration sites.
}
\examples{
\dontrun{ni.scan('P05202')}
}
\references{
Huang et al. Nucleic Acids Res. 2019 47:D298-D308, (PMID: 30418626).
}
\seealso{
meto.scan(), ac.scan(), me.scan(), ub.scan(), su.scan(), gl.scan(), sni.scan(), p.scan(), ptm.scan(), reg.scan(), dis.scan()
}
\author{
Juan Carlos Aledo
}
