% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{pdb.uniprot}
\alias{pdb.uniprot}
\title{Identifier Mapping From PDB to UniProt}
\usage{
pdb.uniprot(id)
}
\arguments{
\item{id}{the identifier to be converted.}
}
\value{
Returns a character string corresponding to the requested identifier.
}
\description{
Mapping between PDB and UniProt protein identifiers.
}
\examples{
\dontrun{pdb.uniprot('3cwm')}
}
\seealso{
id.mapping()
}
\author{
Juan Carlos Aledo
}
