% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_pick_nm.R
\name{ptm_pick_nm}
\alias{ptm_pick_nm}
\title{picks a determination to go with}
\usage{
ptm_pick_nm(ml)
}
\arguments{
\item{ml}{data frame the ptm masterlist}
}
\value{

}
\description{
either the Final determination if there is one of the determindation in the field if there is nothing
}
