% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_to_bold.R
\name{ptm_to_bold}
\alias{ptm_to_bold}
\title{Converts data from the PTM masterlist to the bold template}
\usage{
ptm_to_bold(ptm, filename)
}
\arguments{
\item{ptm}{a list of ptm numbers}

\item{filename}{character - file name}
}
\value{

}
\description{
Converts data from the PTM masterlist to the bold template
}
\examples{
#ptm_to_bold(c("111", "222", "333"), tempfile("savefile"))
}
