% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_to_herbarium.R
\name{ptm_to_herbarium}
\alias{ptm_to_herbarium}
\title{convert PTM master list metadata to UBC herbarium list}
\usage{
ptm_to_herbarium(ptm)
}
\arguments{
\item{ptm}{list ptm numbers character)}
}
\value{

}
\description{
convert PTM master list metadata to UBC herbarium list
}
\examples{
ptm_to_herbarium(c("120","230","300"))
}
