% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{search.go}
\alias{search.go}
\title{Search a Simple User Query}
\usage{
search.go(query)
}
\arguments{
\item{query}{character string defining the query.}
}
\value{
Returns a dataframe containing the GO IDs found associated to the query, as well as other information related to these terms.
}
\description{
Searches a simple user query.
}
\examples{
\dontrun{search.go('oxidative stress')}
}
\references{
Rhee et al. (2008) Nature Reviews Genetics 9:509–515.
}
\seealso{
term.go(), get.go(), bg.go(), hdfisher.go(), gorilla(), net.go()
}
\author{
Juan Carlos Aledo
}
