% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ptm.R
\name{search_ptm}
\alias{search_ptm}
\title{Searches for species in PTM masterlist
looks in the Final determination and Determination in the field}
\usage{
search_ptm(name)
}
\arguments{
\item{name}{scientific name of seaweed}
}
\value{
dataframe
}
\description{
Searches for species in PTM masterlist
looks in the Final determination and Determination in the field
}
\examples{
search_ptm("Ulva lactuca")
}
