% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Gets the higher taxonomy and authorship of the species using the Worms database}
\usage{
taxonomy(sciname)
}
\arguments{
\item{sciname}{chr or list}
}
\value{
dataframe
}
\description{
Gets the higher taxonomy and authorship of the species using the Worms database
}
\examples{
taxonomy("Ulva linza")
taxonomy("Mastocarpus")

# A list of species
taxonomy(c("Ulva linza", "Mastocarpus"))

}
