% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{uniprot2pdb}
\alias{uniprot2pdb}
\title{Return the PDB and Chain IDs of the Provided UniProt Protein}
\usage{
uniprot2pdb(up_id)
}
\arguments{
\item{up_id}{the UniProt ID.}
}
\value{
The function returns a dataframe with info about the PDB related to the protein of interest.
}
\description{
Returns the PDB and chain IDs of the provided protein.
}
\examples{
\dontrun{uniprot2pdb("P04406")}
}
\seealso{
pdb2uniprot(), id.mapping()
}
\author{
Juan Carlos Aledo
}
