% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbarium_template.R
\name{herbarium_template}
\alias{herbarium_template}
\title{Download herbarium template from website}
\usage{
herbarium_template()
}
\value{

}
\description{
Download herbarium template from website
}
\examples{
 #herbarium_template()
}
