% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_to_bold.R
\name{ptm_to_bold}
\alias{ptm_to_bold}
\title{Converts data from the PTM masterlist to the bold template}
\usage{
ptm_to_bold(ptm, filename = "bold.xlsx")
}
\arguments{
\item{ptm}{a list of ptm numbers}

\item{filename}{character - file name, ends with .xlsx}
}
\value{

}
\description{
please review the resulting file before submitting
}
\examples{
ptm_to_bold(c("180", "192", "31"), tempfile("savefile.xlsx"))
}
