% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_to_herbarium.R
\name{ptm_to_herbarium}
\alias{ptm_to_herbarium}
\title{convert PTM master list metadata to UBC herbarium list}
\usage{
ptm_to_herbarium(ptm, filename = "herbarium.csv")
}
\arguments{
\item{ptm}{list ptm numbers (character)}

\item{filename}{name of the file to save to a csv file}
}
\value{

}
\description{
convert PTM master list metadata to UBC herbarium list
}
